refactor(k8s): split health.go grab-bag into focused files
health.go mixed three concerns. Split (same package, no caller changes):
- health.go keeps pod/node health CLASSIFICATION (ClassifyPodHealth,
  ClassifyNodeHealth, PodProblemReason, restart/crashloop helpers).
- detect_node.go: DetectNodeProblems + DetectVersionSkew (joins the detect_ family).
- detect_workload.go: DetectHPAProblems + DetectCronJobProblems.
- format.go: FormatAge/Truncate/ParseCPUToMillis/ParseMemoryToBytes — generic
  display/quantity helpers used in 8 files, which never belonged in health.go.
Test files split to match.
