+++ internal/k8s/health.go
@@ -1,12 +1,8 @@
 package k8s
 
 import (
-	"fmt"
-	"strings"
 	"time"
 
-	autoscalingv2 "k8s.io/api/autoscaling/v2"
-	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 )
 
@@ -362,325 +358,4 @@ func ClassifyNodeHealth(node *corev1.Node) NodeHealth {
 	}
 
 	return h
-}
-
-// NodeProblem describes a detected problem on a node.
-type NodeProblem struct {
-	NodeName string `json:"nodeName"`
-	Problem  string `json:"problem"`
-	Reason   string `json:"reason,omitempty"`
-	Severity string `json:"severity"` // "critical", "high", or "medium"
-}
-
-// DetectNodeProblems scans nodes for NotReady, Cordoned, and pressure conditions.
-func DetectNodeProblems(nodes []*corev1.Node) []NodeProblem {
-	var problems []NodeProblem
-
-	for _, node := range nodes {
-		h := ClassifyNodeHealth(node)
-
-		if !h.Ready {
-			reason := "NotReady"
-			if h.Reason != "" {
-				reason = h.Reason
-			}
-			problems = append(problems, NodeProblem{
-				NodeName: node.Name,
-				Problem:  "NotReady",
-				Reason:   reason,
-				Severity: "critical",
-			})
-		} else if h.Unschedulable {
-			problems = append(problems, NodeProblem{
-				NodeName: node.Name,
-				Problem:  "Cordoned",
-				Reason:   "SchedulingDisabled",
-				Severity: "medium",
-			})
-		}
-
-		for _, pressure := range h.Pressures {
-			problems = append(problems, NodeProblem{
-				NodeName: node.Name,
-				Problem:  pressure,
-				Reason:   pressure,
-				Severity: "critical",
-			})
-		}
-	}
-
-	return problems
-}
-
-// VersionSkew describes a detected minor version skew across cluster nodes.
-type VersionSkew struct {
-	Versions   map[string][]string `json:"versions"` // minor version -> node names
-	MinVersion string              `json:"minVersion"`
-	MaxVersion string              `json:"maxVersion"`
-}
-
-// DetectVersionSkew checks for minor version differences across nodes.
-// Returns nil if all nodes are on the same minor version (patch-only differences are normal).
-func DetectVersionSkew(nodes []*corev1.Node) *VersionSkew {
-	if len(nodes) == 0 {
-		return nil
-	}
-
-	versions := make(map[string][]string) // minor version -> node names
-	for _, node := range nodes {
-		ver := node.Status.NodeInfo.KubeletVersion
-		minor := extractMinorVersion(ver)
-		if minor == "" {
-			continue
-		}
-		versions[minor] = append(versions[minor], node.Name)
-	}
-
-	if len(versions) <= 1 {
-		return nil
-	}
-
-	// Find min and max versions
-	var minV, maxV string
-	for v := range versions {
-		if minV == "" || v < minV {
-			minV = v
-		}
-		if maxV == "" || v > maxV {
-			maxV = v
-		}
-	}
-
-	return &VersionSkew{
-		Versions:   versions,
-		MinVersion: minV,
-		MaxVersion: maxV,
-	}
-}
-
-// extractMinorVersion extracts "v1.28" from "v1.28.3" or "1.28" from "1.28.3".
-func extractMinorVersion(version string) string {
-	version = strings.TrimPrefix(version, "v")
-	parts := strings.SplitN(version, ".", 3)
-	if len(parts) < 2 {
-		return ""
-	}
-	return parts[0] + "." + parts[1]
-}
-
-// FormatAge formats a duration into a human-readable age string (e.g., "5d", "3h").
-func FormatAge(d time.Duration) string {
-	if d < time.Minute {
-		return fmt.Sprintf("%ds", int(d.Seconds()))
-	}
-	if d < time.Hour {
-		return fmt.Sprintf("%dm", int(d.Minutes()))
-	}
-	if d < 24*time.Hour {
-		return fmt.Sprintf("%dh", int(d.Hours()))
-	}
-	return fmt.Sprintf("%dd", int(d.Hours()/24))
-}
-
-// Truncate trims a string to maxLen characters, appending "..." if truncated.
-func Truncate(s string, maxLen int) string {
-	s = strings.TrimSpace(s)
-	if len(s) <= maxLen {
-		return s
-	}
-	return s[:maxLen-3] + "..."
-}
-
-// HPAProblem describes a detected issue with an HPA.
-type HPAProblem struct {
-	Name      string
-	Namespace string
-	Problem   string // "maxed"
-	Reason    string
-}
-
-// DetectHPAProblems finds HPAs that have hit their replica ceiling OR that
-// cannot scale because the autoscaler can't fetch metrics. The latter is
-// the silent-broken-HPA case: spec is valid, target exists, but
-// status.conditions[?type=ScalingActive].status=False means the controller
-// gave up — metrics-server unavailable, broken adapter, missing resource
-// requests on target pods, etc. K8s autoscaler condition reasons are
-// stable across versions (FailedGetResourceMetric / FailedGetScale /
-// FailedGetExternalMetric / FailedGetObjectMetric).
-func DetectHPAProblems(hpas []*autoscalingv2.HorizontalPodAutoscaler) []HPAProblem {
-	var problems []HPAProblem
-	for _, hpa := range hpas {
-		// "maxed" — at replica ceiling and wanting more.
-		if hpa.Spec.MaxReplicas > 0 && hpa.Status.CurrentReplicas >= hpa.Spec.MaxReplicas && hpa.Status.DesiredReplicas >= hpa.Spec.MaxReplicas {
-			problems = append(problems, HPAProblem{
-				Name:      hpa.Name,
-				Namespace: hpa.Namespace,
-				Problem:   "maxed",
-				Reason:    fmt.Sprintf("%d/%d replicas (wants %d)", hpa.Status.CurrentReplicas, hpa.Spec.MaxReplicas, hpa.Status.DesiredReplicas),
-			})
-		}
-		// "cannot scale" — the autoscaler controller reports it can't get
-		// metrics or scale calls are failing. Emitted as a separate problem
-		// so the maxed-check above isn't masked by an unrelated metrics
-		// outage on the same HPA.
-		for _, cond := range hpa.Status.Conditions {
-			if cond.Type == autoscalingv2.ScalingActive && cond.Status == corev1.ConditionFalse {
-				reason := cond.Reason
-				if reason == "" {
-					reason = "ScalingActive=False"
-				}
-				msg := cond.Message
-				if msg == "" {
-					msg = "HPA controller reports it cannot scale this workload"
-				}
-				problems = append(problems, HPAProblem{
-					Name:      hpa.Name,
-					Namespace: hpa.Namespace,
-					Problem:   "cannot-scale",
-					Reason:    fmt.Sprintf("%s: %s", reason, msg),
-				})
-				break
-			}
-		}
-	}
-	return problems
-}
-
-// CronJobProblem describes a detected issue with a CronJob.
-type CronJobProblem struct {
-	Name      string
-	Namespace string
-	Problem   string // "stale" or "never-scheduled"
-	Reason    string
-}
-
-// estimateCronMinInterval returns a coarse lower bound on the time between runs
-// of a standard 5-field cron schedule (minute hour dom month dow), plus the
-// common @-macros. It is deliberately approximate — its only job is to keep
-// DetectCronJobProblems from flagging a rare-cadence job (weekly / monthly /
-// quarterly) as "stale" against a flat daily threshold. ok=false for schedules
-// it can't parse; the caller then falls back to the flat threshold.
-func estimateCronMinInterval(schedule string) (time.Duration, bool) {
-	const day = 24 * time.Hour
-	s := strings.TrimSpace(schedule)
-	switch s {
-	case "@yearly", "@annually":
-		return 365 * day, true
-	case "@monthly":
-		return 28 * day, true
-	case "@weekly":
-		return 7 * day, true
-	case "@daily", "@midnight":
-		return day, true
-	case "@hourly":
-		return time.Hour, true
-	}
-	fields := strings.Fields(s)
-	if len(fields) != 5 {
-		return 0, false
-	}
-	hour, dom, month, dow := fields[1], fields[2], fields[3], fields[4]
-	switch {
-	case month != "*":
-		// Constrained to certain months → at most monthly, often far less.
-		return 28 * day, true
-	case dom != "*":
-		// Specific day(s)-of-month → monthly cadence.
-		return 28 * day, true
-	case dow != "*":
-		// Specific day(s)-of-week → weekly is the conservative lower bound.
-		return 7 * day, true
-	case hour != "*" && !strings.HasPrefix(hour, "*/"):
-		// Specific hour(s) each day → daily.
-		return day, true
-	default:
-		// Intra-day cadence (every minute / */n minutes or hours).
-		return time.Hour, true
-	}
-}
-
-// DetectCronJobProblems finds non-suspended CronJobs that haven't run recently.
-func DetectCronJobProblems(cronjobs []*batchv1.CronJob) []CronJobProblem {
-	var problems []CronJobProblem
-	now := time.Now()
-	for _, cj := range cronjobs {
-		if cj.Spec.Suspend != nil && *cj.Spec.Suspend {
-			continue
-		}
-		// Staleness is relative to the schedule's cadence, not a flat day: a
-		// quarterly job that ran on schedule 29 days ago is healthy, not stale.
-		// Floor at 24h so frequent jobs keep the original sensitivity.
-		threshold := 24 * time.Hour
-		if interval, ok := estimateCronMinInterval(cj.Spec.Schedule); ok {
-			if grace := interval + interval/2; grace > threshold {
-				threshold = grace
-			}
-		}
-		if cj.Status.LastScheduleTime != nil {
-			sinceLast := now.Sub(cj.Status.LastScheduleTime.Time)
-			if sinceLast > threshold {
-				problems = append(problems, CronJobProblem{
-					Name:      cj.Name,
-					Namespace: cj.Namespace,
-					Problem:   "stale",
-					Reason:    fmt.Sprintf("last run %dh ago", int(sinceLast.Hours())),
-				})
-			}
-		} else if now.Sub(cj.CreationTimestamp.Time) > threshold {
-			problems = append(problems, CronJobProblem{
-				Name:      cj.Name,
-				Namespace: cj.Namespace,
-				Problem:   "never-scheduled",
-				Reason:    "created but never ran",
-			})
-		}
-	}
-	return problems
-}
-
-// ParseCPUToMillis parses CPU quantity strings like "250m", "1", "500n".
-func ParseCPUToMillis(s string) int64 {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return 0
-	}
-	if before, ok := strings.CutSuffix(s, "n"); ok {
-		var val int64
-		fmt.Sscanf(before, "%d", &val)
-		return val / 1000000
-	}
-	if before, ok := strings.CutSuffix(s, "m"); ok {
-		var val int64
-		fmt.Sscanf(before, "%d", &val)
-		return val
-	}
-	var val int64
-	fmt.Sscanf(s, "%d", &val)
-	return val * 1000
-}
-
-// ParseMemoryToBytes parses memory quantity strings like "1024Ki", "256Mi", "1Gi".
-func ParseMemoryToBytes(s string) int64 {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return 0
-	}
-	if before, ok := strings.CutSuffix(s, "Ki"); ok {
-		var val int64
-		fmt.Sscanf(before, "%d", &val)
-		return val * 1024
-	}
-	if before, ok := strings.CutSuffix(s, "Mi"); ok {
-		var val int64
-		fmt.Sscanf(before, "%d", &val)
-		return val * 1024 * 1024
-	}
-	if before, ok := strings.CutSuffix(s, "Gi"); ok {
-		var val int64
-		fmt.Sscanf(before, "%d", &val)
-		return val * 1024 * 1024 * 1024
-	}
-	var val int64
-	fmt.Sscanf(s, "%d", &val)
-	return val
 }
